Fix : [BE] tts 디폴트 이미지 설정

See merge request s11-final/S11P31D105!165

+++ backend/src/main/java/com/checkitout/ijoa/TTS/service/TTSService.java
@@ -26,6 +26,7 @@ import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
 
@@ -75,7 +76,14 @@ public class TTSService {
             throw new CustomException(ErrorCode.TTS_LIMIT_EXCEEDED);
         }
 
-        String url = fileService.saveProfileImage(requestDto.getImage());
+        String url;
+        MultipartFile image = requestDto.getImage();
+        if(requestDto.getImage()!=null){
+            url = fileService.saveProfileImage(requestDto.getImage());
+
+        }else{
+            url="https://checkitout-bucket.s3.ap-northeast-2.amazonaws.com/ttsprofile/default.png";
+        }
 
         TTS newTTS = TTSProfileRequestDto.of(requestDto,url,user);
 
@@ -94,9 +102,12 @@ public class TTSService {
         checkUser(deleteTTS,user.getId());
 
         //s3 파일 삭제
-        // 프로필 이미지 삭제
-        String key = getKeyFromUrl(deleteTTS.getImage());
-        fileService.deleteFile(key);
+        // 프로필 이미지 디폴트 아니면 삭제
+        if(!"https://checkitout-bucket.s3.ap-northeast-2.amazonaws.com/ttsprofile/default.png".equals(deleteTTS.getImage())){
+            String key = getKeyFromUrl(deleteTTS.getImage());
+            fileService.deleteFile(key);
+
+        }
         // 학습 데이터 삭제
         List<TrainAudio> trainAudios = trainAudioRepository.findByTtsId(deleteTTS.getId());
         if(trainAudios!=null){
@@ -126,13 +137,18 @@ public class TTSService {
 
         TTS updateTTS = ttsRepository.findById(ttsId).orElseThrow(()-> new CustomException(ErrorCode.TTS_NOT_FOUND));
         checkUser(updateTTS,user.getId());
+        String url;
 
         if(requestDto.getImage() != null && !requestDto.getImage().isEmpty()){
             //기존 이미지 s3삭제
-            String key = getKeyFromUrl(updateTTS.getImage());
-            fileService.deleteFile(key);
 
-            String url = fileService.saveProfileImage(requestDto.getImage());
+            if(!"https://checkitout-bucket.s3.ap-northeast-2.amazonaws.com/ttsprofile/default.png".equals(updateTTS.getImage())){
+                String key = getKeyFromUrl(updateTTS.getImage());
+                fileService.deleteFile(key);
+
+            }
+
+            url = fileService.saveProfileImage(requestDto.getImage());
             updateTTS.setImage(url);
         }
